minor: Handle mismatches between oplog + known ops
The current implementation handles mismatches between the op-log and
known set of operations quite badly.

It is required that the current op-log head points to a known and
explicitly provided operation.

This can cause issues in the following scenarios:

+ Working accross two different branches and applied operations from one
  branch not existing in the second.
+ Deleting applied operations because it is no longer desireable to keep
  the code around.

Another issue exists as a result of this approach; information regarding
the order in which operations were actually applied is lost. This may be
distinct from the order in which operations were provided.

When executing in the :down direction it would be safer to execute in
the reverse order to which operations were applied.

This commit resolves this issue by:
+ No longer requiring operations that have been applied to be provided at all.
+ Re-ordering operations based on the order they appear in the op-log.