major: Rename term :migration[s] to :operation[s]
While this tool is primarily intended to be used for running migrations,
it can in fact be used for running any kind of once off operation. For
example this can be used for seeding data in dev.

In order to encourage its broader use any instances of the term
'migration' has been renamed in this commit to 'operation'.

This change had already started to be implemented in previous commits,
but this commit changes the public API to follow suite.

This will make it seem less weird to use a migration tool for other
use-cases.